src/moomoo_mcp/services/base_service.py
from moomoo import OpenQuoteContext, RET_OK

class MoomooService:
    """Service to manage Moomoo API connections."""
    
    def __init__(self, host: str = '127.0.0.1', port: int = 11111):
        self.host = host
        self.port = port
        self.quote_ctx: OpenQuoteContext | None = None

    def connect(self) -> None:
        """Initialize connection to OpenD."""
        # OpenQuoteContext connects on initialization
        self.quote_ctx = OpenQuoteContext(host=self.host, port=self.port)
        
    def close(self) -> None:
        """Close connection."""
        if self.quote_ctx:
            self.quote_ctx.close()
            self.quote_ctx = None

    def check_health(self) -> dict[str, str]:
        """Check connection health."""
        if not self.quote_ctx:
            return {"status": "disconnected", "error": "Context not initialized"}
        
        # Simple check: get_global_state or similar if available, or just heuristic
        # moomoo-api doesn't have a direct 'ping', but we can check if socket is alive 
        # or try a lightweight call.
        # For now, if we instantiated it and it didn't raise, we assume connected 
        # (though OpenQuoteContext is sync and might strict connect).
        
        # Let's try to get global state if possible, or just return connected.
        # Actually moomoo-api often logs errors if connection fails.
        # We can implement a try-catch around a simple call later.
        
        # Returning connected for now as a basic check.
        # Real implementation would call something lightweight.
        return {"status": "connected", "host": f"{self.host}:{self.port}"}
