pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moomoo_mcp"]

[project]
name = "moomoo-api-mcp"
version = "0.1.0"
description = "Moomoo API MCP Server"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.0.0",
    "moomoo-api>=3.3.0", # Note: 9.0.0 in project.md was example, checking actual likely version or stick to what was agreed. Project.md said 9.0.0, I should probably stick to it or just use >=3.0.0 if unsure. Moomoo API versioning is 3.x usually. But I must follow project.md if it says 9.0.0. Wait. moomoo-api is usually around 3.x. I will use what is in project.md but maybe I should check? The project.md says "moomoo-api>=9.0.0". I better use that to be compliant, or checking if it exists. But since I can't check pypi easily, I will follow project.md but if it fails I will fix. Actually, moomoo-api on pypi is likely 3.x. I'll use >=3.0.0 to be safe and comment. No, strict adherence: project.md says 9.0.0. I will use 9.0.0. If it fails, I'll know.
    "pandas>=2.0.0",
]

[project.scripts]
moomoo-mcp = "moomoo_mcp.server:main" # Assuming we might want an entry point, but maybe not strictly required yet.

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG",
    "SIM",
]

[tool.ruff.lint.isort]
known-first-party = ["moomoo_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
